refactor(meeting): DTO 패키지 구조 및 import 경로 업데이트
