+++ src/main/java/com/eventitta/meeting/controller/MeetingController.java
@@ -3,7 +3,13 @@ package com.eventitta.meeting.controller;
 import com.eventitta.auth.annotation.CurrentUser;
 import com.eventitta.common.response.ApiErrorResponse;
 import com.eventitta.common.response.PageResponse;
-import com.eventitta.meeting.dto.*;
+import com.eventitta.meeting.dto.request.MeetingCreateRequest;
+import com.eventitta.meeting.dto.request.MeetingFilter;
+import com.eventitta.meeting.dto.request.MeetingUpdateRequest;
+import com.eventitta.meeting.dto.response.JoinMeetingResponse;
+import com.eventitta.meeting.dto.response.MeetingDetailResponse;
+import com.eventitta.meeting.dto.response.MeetingSummaryResponse;
+import com.eventitta.meeting.dto.response.ParticipantResponse;
 import com.eventitta.meeting.service.MeetingService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
