+++ src/main/java/com/eventitta/meeting/repository/MeetingRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.eventitta.meeting.repository;
 
-import com.eventitta.meeting.dto.MeetingFilter;
-import com.eventitta.meeting.dto.MeetingSummaryResponse;
+import com.eventitta.meeting.dto.request.MeetingFilter;
+import com.eventitta.meeting.dto.response.MeetingSummaryResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
