+++ src/test/java/com/eventitta/meeting/controller/MeetingControllerTest.java
@@ -4,9 +4,9 @@ import com.eventitta.WithMockCustomUser;
 import com.eventitta.common.config.SecurityConfig;
 import com.eventitta.meeting.domain.MeetingStatus;
 import com.eventitta.meeting.domain.ParticipantStatus;
-import com.eventitta.meeting.dto.MeetingCreateRequest;
-import com.eventitta.meeting.dto.MeetingUpdateRequest;
-import com.eventitta.meeting.dto.ParticipantResponse;
+import com.eventitta.meeting.dto.request.MeetingCreateRequest;
+import com.eventitta.meeting.dto.request.MeetingUpdateRequest;
+import com.eventitta.meeting.dto.response.ParticipantResponse;
 import com.eventitta.meeting.service.MeetingService;
 import com.eventitta.auth.jwt.JwtTokenProvider;
 import com.eventitta.auth.service.CustomUserDetailsService;
